Changed Spark values in contracts based on a more correct understanding of how ERC20 tokens work. They define decimal places, but only store integers. So you define value in the smallest possible unit. In our case, 1e-18. This makes one Spark worth 1e18 such units.

+++ contracts/SparkAuction.sol
@@ -104,7 +104,7 @@ contract SparkAuction is Ownable
     StrikePriceSet(_price);
   }
 
-  function acceptBid(uint256 _price, uint16 _quantity, uint8 _v, bytes32 _r,
+  function acceptBid(uint256 _price, uint256 _quantity, uint8 _v, bytes32 _r,
                      bytes32 _s)
     external
     afterSale
